requirements.txt
# Core
numpy==2.4.3
pandas==2.3.3

# ML
scikit-learn==1.8.0
tensorflow==2.21.0
joblib==1.5.3

# API
flask==3.1.3
requests==2.32.5
python-dotenv==1.2.2

# Kafka
kafka-python==2.3.0

# Dashboard
streamlit==1.55.0
plotly==6.6.0

# Visualization
matplotlib==3.10.8
seaborn==0.13.2

# Testing
pytest==9.0.2